refactor(buildFromDefinition) move classes into files
